+++ src/assets/images/index.tsx
@@ -1,4 +1,5 @@
 import CardChip from './cardChip.png';
+import Check from './check.jpg';
 import EtcCard from './etc.jpg';
 import MasterCard from './mastercard.jpg';
 import VisaCard from './visa.jpg';
@@ -8,6 +9,7 @@ const IMAGES = {
   etcCard: EtcCard,
   masterCard: MasterCard,
   visaCard: VisaCard,
+  check: Check,
 };
 
 export default IMAGES;
